Heh--made a big frigging mess in data.py, player.py and app.py and stdcmds.py.
I'm about half way done two different things: add_command and import module.
Then folks can build stuff in other modules and import it on the fly.  In
addition, it'll add commands.  Also, we'll push our commands to the stdcmds
file or something like that.  (Maybe to the bottom of player.py but that
is such a lousy naming scheme).