split the session class into two classes, backwardly compatible. session is now a base class and UserSession inherits from it.  session contains connection state and basic settings needed by all types of sessions.  UserSession contains state and methods which make up the Lyntin core functionality.  This decouples the ideas of "connection to a server" and "Lyntin state," which should make it easier to make other types of sessions.
Added the handler class, used for intercepting server output at a low
level.  The idea is that sessions have many handlers, which serve to
interpret mud output and filter it for the next handler.  The only
current handler is AppHandler, which calls out to the regular lyntin
output handler and doesn't filter the output at all.

added directory libcore, where we can put utility modules that can be
loaded early and don't depend on the core.

fixed buglet in aborting due to errors.

added debugging flag to lyntin.py so you get full tracebacks when
debugging early initialization.