Moved help files to a .help extension.  I should prolly change this to .hlp--but I'll do that later.
Also changed Putline, PutUntouchedLine, PutReallyUntouchedLine and other
such unclear print-type functions to PutError, PutMessage, PutUserInput
and PutRaw.  Changed this across the whole code base.

In the process I re-wrote some of the output messages and fixed a few
bugs.

I also changed ParseFile so that it executes dispatch_command on
all the lines in the file rather than having its own command interpreter
which seemed kind of silly.  This causes the unhappy effect of it being
extremely verbose.  I'm going to think about how to make it quieter.
Perhaps create a switch to quell lyntin messages?  Is that what
#verbose does anyhow?

All this and more in the next commit!