Changed so that argument parsing only happens on the arguments and not the full command line. This way the noparsing option will produce a dictionary with values {"command":"showme", "input":"how to make a cake"} where previously it would have produced an argument dictionary of the form {"input":"showme how to make a cake"}
session.py takes care of pulling out the command name from the line and passing it to
argparser, then populating the "command" element of the argument dictionary.

engine.py no longer has the hack ""command;string" + arguments" addition to the argspec.

nop, showme, and raw were all changed to take advantage of this.