minor tweaks.
added a tag argument to handleUserData and write_socket.  This argument
simply gets sent along and passed out as part of the tuple in
to_mud_hook.  Otherwise it is totally ignored.

Changed session.py to set dict["command"] to be the full command name, not
what the user entered.