Aletered behaviour of index collector arguments, and removed the : (which didn't work anyway) meaning from argspecs.
Now an index collector can appear anywhere before the named
collector.  Arguments after the index collector must be set by name
only.  (They should have defaults, otherwise you'll always be specifying
that argument name, but this isn't enforced.)

So an arspec like "options* quiet:boolean=true" will take any number of
options, and you'll have to set quiet with a "quiet=blah" entry rather
than with any arguments.

Also changed it so that when there is an index collector named arguments
can be intersperced with unnamed arguments.  (This can still make things
confusing when a named argument is early in a list with multiple index
arguments, but it allows for the quiet=blah to be anywhere in the above
example.)