Incorporated some of sebastians changes and did some cleanup.
It is very important that speedwalk.filter checks that it is not
processing things that are speedwalk aliases.  the check that
text not in self._dirs.values() ensures that we avoid most infinite
recursions through speedwalking alone.

self._aliases is used to keep track of the value of
self._dirs.values() without having to recompute

I also figured since we're storing self._dirs.values() so that we don't
have to recompute it, I would also store self._dirs.keys() in
self._dirs_available, so it doesn't have to be recomputed with each call
to filter.