Complete overhaul of the help system number 2 based on conversations with Josh on the million things he wanted the help system to do.  It now stores things dynamically and hierarchically with several methods of adding help.
For commands that are bound to Lyntin, the help system will grab the
syntaxline generated by the argparser (if it has argparser handled
arguments) and the __doc__ attribute of the function object.

For files ending in .tpc located in the 'help' package, these are now
loaded at Lyntin startup with everything else.

Additionally, you can add help topics manually through
'exported.add_help'.

Help texts are now categorized hierarchically.  A fully qualified
help text name is delimited by periods (".") and ends with the name
of the help text.  For example:

   general.command

is the command.tpc file.

You can specify the categories hierarchy that your text sits at
as an argument to exported.add_help.  Additionally, the helpmanager
will examine the first line of your help text for a category line
there.  The category line must be the very first line and must
start with "category: " and have a period delimited list of
categories the text is categorized under.

This is a first revision and there's some things missing.  Enhancements
to this basic functionality will come later as we agree that this is
the direction we want to take with this.