Added in searching features to helpmanager.getHelp.  Now if the appropriate help node is not found by interpreting the input as a fully qualified name it searches for all nodes with matching names.  If a single result is found it is returned as if it had been asked for directly, otherwise text listing the found help items is returned.
Also changed it so that addHelp looks for the categories: line at the end
of the help text, not at the start.  This lets us add on the sytax: line
to the help text in addCommand,a nd it doesn't have to be pulled out by
the #help_cmd.  Changed the topics correspondingly.