Rearranged and cleaned up and extended the argparser type system.
Added three new types:

choice  ---  allows for values from a pipe-delimitted list of
choices.  will automatically expand uniquely specified values that match
one of the possible values.

time, timespan --- accept time delimitters.  time accepts a specific time
of day and returns the nearest occurence in the future, timespan accepts a
timespan with hours/minutes/seconds specified with h/m/s, or assumes the
number is seconds if no character is specified.  time also accepts a
timespan and uses it as a time to add to time.time().