MASSIVE CHANGES!
This includes the complete manager overhaul.  Managers are now just
Lyntin modules--allowing us to import them and change them and prevent
them from being loaded at Lyntin startup just like any other Lyntin
module.  This gives us crazy amounts of power and everything.

At some point I'll update the HACKER doc--it needs it.

Also included is a complete re-write as to #write session persistence.
Register with the write_hook and you too can persist stuff with #write!

Also included is a series of minor fixes here and there to #info
and such.

And that's not all!  I also broke builtin variables and probably caused
a bunch of other issues which will take time to discover and fix and
such.