Changes and updates to the HACKING guide (still a work in progress). Changed the write_hook so that it passes the file object as well as a quiet argument which indicates whether to persist things in a quiet fashion (if such a thing is possible).  So persisting an alias would add:
   #alias {a} {b}

to the file, but persisting it quietly would add:

   #alias {a} {b} quiet={true}

See the difference?  I have yet to cycle this into all of the managers,
but the ground-work is there in both the write_cmd and in the write_hook.
Minor changes to the config_cmd so that it accepts the quiet argument
as well.