All kinds of wild and exciting changes in the commit.  Some of them may cause you much STRESS because I fixed the color system and it's not backwards compatible with the old system (which was borked).
The new system correctly handles bold, underline, blink, and reverse.
These are all now supported in highlight and the textui (in a terminal
that supports this sort of stuff) will handle all of them.  The tkui
now supports reverse, underline and bold.  Blink doesn't work
in the tkui.  I'm happy to take recommendations on how to handle this.

I created a BellEvent, bell_hook, and the underlying stuff in net.py
to pick out the bell and kick off the BellEvent which spams the
bell_hook with the session in question.  Implement a bell in your ui
if you so desire!

I also fixed the color system to use a series of constants for all
kinds of stuff and changed the rest of the code-base to use those
constants.  That'll make future color changes easier to handle.

And I added support for more of the xterm/vt220 color handling stuff.

And to finish it off, I added telnet control code conversation logging.
If you want to see what's going on, it's in the SocketCommunicator's
_controllog member.  It keeps track of what we received and what we
sent and in order.  Makes debugging weird telnet control code conversation
issues much much much easier.

I think that's it.