This is the testmodule, tests, and all the benchmarks.
To use:

   #load modules.testsystem.testmodule

and then:

   #runtest

If there is no benchmark for a given test (which is just a command
file), then it'll create one.  If there is a benchmark, it'll diff the
results it just got with those in the benchmark and let you know whether
the test passed or failed.