Removed the BellEvent (we now spam the bell_hook with the SpamEvent).
Added prompt detection which will spam a prompt_hook.  We detect prompts
by looking for IAC+GA or IAC+TELOPT_EOR.  There are provisions for looking
for a text string as well, but the only way to trigger prompt string
detection is via a plugin that calls _buildPromptRegex on the
SocketCommunicator of the Session.