This is a monster set of changes all kind of rolled into one massive checking.  Sorry bout that.  They sort of compounded themselves.
Change 1:
I'm overhauling the configuration stuff.  I switched commandchar so
that it's in the Configuration Manager.  I'm still adjusting the API.
Getting there...  Sorry for putting everyone in temporary limbo between
two configuration architectures.

Big Change 2:
I went through and fixed how we persist command stuff via the write
command because the old way was really crappy code.  In the new way,
all the things that are registered with teh write_hook should return
a list of strings representing the things they want to persist.  For
example, if I wanted the Alias Manager to persist three aliases, it
would return:

   ["alias {s %1} {smile %1}",
    "alias {bi} {buy insecticide}",
    "alias {se} {southeast}"]

Then the write_cmd code in tintincmds.py will take all those
strings, prepend the commandchar to beginning and append the \n to
the end of every line.  And then it worries about writing it to the
file.  No more passing the file pointer around.  Ick.

Change 3:
While doing that, I cleaned up a lot of the getInfo methods floating
around.  I also fixed the tag stuff in action--the disabled tags
needed their own method--otherwise they get persisted with the rest
of the actions and I'm pretty sure we don't want that to happen.


Everything is mostly tested, but not incredibly so.  I'm still going
through and making changes for the configuration manager.  Some things
are a little bit of a squirrely mess.